backend/requirements.txt
fastapi
uvicorn
python-multipart
scikit-learn
pydantic-settings
PyPDF2
openai
python-dotenv